Herraria.Level

Description

Module with generation of the level

Documentation

data Block

Constructors

Air

Ground

Bedrock

newtype Chunk a

Chunk of Blocks Dimensions: chunkHeight by chunkWidth

Constructors

Chunk

Fields

getChunk :: Vector (Vector a)

data Level

Level is an infinite list of chunks with focus on current player position

Constructors

Level

Fields

leftChunks :: [Chunk Block]

curChunk :: Chunk Block

rightChunks :: [Chunk Block]

curIndex :: Int

chunkHeight :: Int

Default chunk height

chunkWidth :: Int

Default chunk width

putBlock :: (Int, Int) -> Block -> Chunk Block -> Chunk Block

Put given block to given position in Chunk

imapChunk :: (Int -> Int -> a -> b) -> Chunk a -> Chunk b

Apply a function to every element of a chunk and its position.

chunkToLists :: Chunk a -> [[a]]

Convert Chunk to List of Lists.

chunkFromLists :: [[a]] -> Chunk a

Convert List of Lists to Chunk.

moveToLeft :: Level -> Level

Load next left chunk.

moveToRight :: Level -> Level

Load next right chunk.

blockToRigidBody

Arguments

:: Float
Ofset of chunk regarding to zero chunk.

-> Int
X position of block in chunk.

-> Int
Y position of block in chunk.

-> Block
Type of block.

-> Maybe (RigidBody Block)
Physical body of the object, if block not equal to Air.

Transform Block to Physical body.

blockToCoords

Arguments

:: Int
X position of block in chunk

-> Int
Y position of block in chunk

-> Block
Type of block

-> (Point, Point)
Coordinates and size

Transform block to coordinates and size

emptyChunk :: Chunk Block

Chuk with all Air

defaultChunk :: Chunk Block

Default chunk with badrock, ground and air

infiniteRandomValues

Arguments

:: (Int -> a)
Function wich will apply to random values

-> StdGen
Initial generator

-> [a]
List of values

Generates infinite list of random values

generateRandomGround

Arguments

:: Int
maximum height

-> StdGen
Random generator

-> Vector (Vector Block)
Sub chunk of ground

Generate absolute random subchunk of ground

generateRandomColumnGround

Arguments

:: Int
Min height

-> Int
Max height

-> StdGen
Random generator

-> Vector (Vector Block)
Subchunk of ground

Generate random column of ground Generate value from minHeight to maxHeight, make column of ground with generated height

generateRecurentGround

Arguments

:: Int
Max height

-> Int
Base value of height

-> Int
Min delta

-> Int
Max delta

-> StdGen
Random generator

-> Vector (Vector Block)
Subchunk of ground

Generate random delta. Height of column of ground equal to height of previous height + delta

randomChunk :: StdGen -> Chunk Block

Generate random chunk with recurent random ground.

defaultLevel :: Level

Constructor of level