Herraria.World

Description

Module responsible for interaction with World

Documentation

data GameState

State of the game

Constructors

GameState

Fields

gamePlayer :: Player

Player

gameLevel :: Level

Current level

gameKeys :: Set Key

Save pressed keys, needed to move player on hold down the key

scaleInput :: Point -> Point

Scale gloss coordinates to game coordinates

shiftInput :: Point -> Point

Move coordinates to the center

mouseToPlayer :: Point -> Point

Mouse coordinates to game coordinates

pointToInt :: (Float, Float) -> (Int, Int)

Floor point

inputToChunkInt :: Float -> Point -> Point -> (Int, Int)

Input coordinates to block in chunk coordinates

updatePressed :: KeyState -> Key -> Set Key -> Set Key

Update pressed keys

blockToPut :: MouseButton -> Block

From mouse button to block

handleWorld :: Event -> GameState -> GameState

Update world

keyToDirection :: Key -> Maybe Direction

Map key to Direction

directionToVec :: Direction -> Point

Map direction to vector

checkCollisionWithLevel

Arguments

:: RigidBody Player
Physical body of player.

-> Level
Level.

-> Bool
True if there are collision, False if there are not.

Check Colision of player with level

notInChunk :: Float -> Point -> Chunk Block -> Maybe Direction

Check either point in chunk or not. Return Nothing if in chunk. Return Just Left if the point left from chunk. Return Just Right if the point right from chunk.

updatePhysics :: Float -> GameState -> GameState

Update Physics of the game. Move player and level if needed Check collisions of player with level

initWorld :: GameState

Default world